+++ src/classes/managers/orders/errors/close-position.error.ts
@@ -1,0 +1,7 @@
+import { cTraderXError } from '../../../models/ctrader-x-error.model';
+
+export class ClosePositionError extends cTraderXError {
+    constructor(error: unknown) {
+        super(`Close position error: ${cTraderXError.getMessageError(error)}`);
+    }
+}