+++ src/classes/managers/orders/errors/get-position-unrealized-pnl.error.ts
@@ -1,0 +1,7 @@
+import { cTraderXError } from '../../../models/ctrader-x-error.model';
+
+export class GetPositionUnrealizedPnLError extends cTraderXError {
+    constructor(error: unknown) {
+        super(`Get position unrealized PnL error: ${cTraderXError.getMessageError(error)}`);
+    }
+}