+++ src/examples/get-position-unrealized-pnl.example.ts
@@ -1,0 +1,14 @@
+import { cTraderX } from '../classes/client';
+
+// Get position unrealized PnL example
+(async () => {
+    const client = new cTraderX();
+
+    await client.connect();
+    const result = await client.orders.getPositionUnrealizedPnL();
+    console.log(result);
+
+    setTimeout(() => {
+        client.disconnect();
+    }, 1000);
+})();